Refactor to use separate components
