+++ planner/src/views/Plan.vue
@@ -2,20 +2,38 @@
     <div class="plan">
         <Navbar></Navbar>
         <br>
-        <Parameters></Parameters>
+        <b-container class="float-left">
+            <b-row>
+                <b-col class="col-3">
+                    <!--b-list group for dive segments-->
+                    <BottomSegments></BottomSegments>
+                </b-col>
+                <b-col class="col-3">
+                    <!--b-list group for deco gases-->
+                    <DecoGases></DecoGases>
+                </b-col>
+                <b-col>
+                    <!--reserved for showing dive results-->
+                    <p>Reserved for showing dive results</p>
+                </b-col>
+            </b-row>
+        </b-container>
     </div>
 </template>
 
 <script lang="ts">
     import Navbar from "@/components/Navbar.vue";
     import {Component, Vue} from "vue-property-decorator";
-    import Parameters from "@/components/plan/Parameters.vue";
+    import BottomSegments from "@/components/plan/BottomSegments.vue";
+    import DecoGases from "@/components/plan/DecoGases.vue";
 
     @Component({
         name: "Plan",
         components: {
+            DecoGases,
+            BottomSegments,
             Navbar,
-            Parameters
+
         }
     })
